# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gamerbuddy API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface SignupBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone: string;
  officialId?: Blob;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  idVerified: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message?: string;
}

export interface Wallets {
  hiringBalance: number;
  earningsBalance: number;
  canWithdraw: boolean;
  canPostRequest: boolean;
}

export interface DepositBody {
  /**
   * @minimum 10.75
   * @maximum 1000
   */
  amount: number;
}

export interface WithdrawBody {
  amount: number;
}

export type GameRequestPlatform =
  (typeof GameRequestPlatform)[keyof typeof GameRequestPlatform];

export const GameRequestPlatform = {
  PC: "PC",
  PlayStation: "PlayStation",
  Xbox: "Xbox",
  Nintendo_Switch: "Nintendo Switch",
  Steam_Deck: "Steam Deck",
  iOS: "iOS",
  Android: "Android",
} as const;

export type GameRequestSkillLevel =
  (typeof GameRequestSkillLevel)[keyof typeof GameRequestSkillLevel];

export const GameRequestSkillLevel = {
  Beginner: "Beginner",
  Intermediate: "Intermediate",
  Expert: "Expert",
  Chill: "Chill",
} as const;

export type GameRequestStatus =
  (typeof GameRequestStatus)[keyof typeof GameRequestStatus];

export const GameRequestStatus = {
  open: "open",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface GameRequest {
  id: number;
  userId: number;
  userName?: string;
  gameName: string;
  platform: GameRequestPlatform;
  skillLevel: GameRequestSkillLevel;
  objectives: string;
  status: GameRequestStatus;
  createdAt: string;
}

export type CreateRequestBodyPlatform =
  (typeof CreateRequestBodyPlatform)[keyof typeof CreateRequestBodyPlatform];

export const CreateRequestBodyPlatform = {
  PC: "PC",
  PlayStation: "PlayStation",
  Xbox: "Xbox",
  Nintendo_Switch: "Nintendo Switch",
  Steam_Deck: "Steam Deck",
  iOS: "iOS",
  Android: "Android",
} as const;

export type CreateRequestBodySkillLevel =
  (typeof CreateRequestBodySkillLevel)[keyof typeof CreateRequestBodySkillLevel];

export const CreateRequestBodySkillLevel = {
  Beginner: "Beginner",
  Intermediate: "Intermediate",
  Expert: "Expert",
  Chill: "Chill",
} as const;

export interface CreateRequestBody {
  gameName: string;
  platform: CreateRequestBodyPlatform;
  skillLevel: CreateRequestBodySkillLevel;
  objectives: string;
}

export interface DashboardSummary {
  user: User;
  wallets: Wallets;
  totalRequestsPosted: number;
  openRequestsCount: number;
  recentRequests: GameRequest[];
}

export type ListRequestsParams = {
  platform?: string;
  skillLevel?: string;
  status?: string;
};