# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gamerbuddy API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User operations
  - name: wallets
    description: Wallet operations
  - name: requests
    description: Game requests operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wallets:
    get:
      operationId: getWallets
      tags: [wallets]
      summary: Get current user's wallets
      responses:
        "200":
          description: Wallet balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallets"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wallets/deposit:
    post:
      operationId: depositHiringWallet
      tags: [wallets]
      summary: Add money to hiring wallet (min 10.75, max 1000)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositBody"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallets"
        "400":
          description: Invalid amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wallets/withdraw:
    post:
      operationId: withdrawEarningsWallet
      tags: [wallets]
      summary: Withdraw from earnings wallet (min balance 100)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawBody"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallets"
        "400":
          description: Invalid amount or insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List all game requests
      parameters:
        - name: platform
          in: query
          schema:
            type: string
        - name: skillLevel
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Post a new game request (requires 10.75 in hiring wallet)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameRequest"
        "400":
          description: Validation error or insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /requests/my:
    get:
      operationId: getMyRequests
      tags: [requests]
      summary: Get current user's posted requests
      responses:
        "200":
          description: List of user's requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameRequest"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a specific request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [users]
      summary: Get dashboard summary for current user
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    SignupBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
        officialId:
          type: string
          format: binary
      required:
        - name
        - email
        - password
        - phone

    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        idVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - idVerified
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user

    Wallets:
      type: object
      properties:
        hiringBalance:
          type: number
        earningsBalance:
          type: number
        canWithdraw:
          type: boolean
        canPostRequest:
          type: boolean
      required:
        - hiringBalance
        - earningsBalance
        - canWithdraw
        - canPostRequest

    DepositBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 10.75
          maximum: 1000
      required:
        - amount

    WithdrawBody:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount

    GameRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        gameName:
          type: string
        platform:
          type: string
          enum: [PC, PlayStation, Xbox, "Nintendo Switch", "Steam Deck", iOS, Android]
        skillLevel:
          type: string
          enum: [Beginner, Intermediate, Expert, Chill]
        objectives:
          type: string
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - gameName
        - platform
        - skillLevel
        - objectives
        - status
        - createdAt

    CreateRequestBody:
      type: object
      properties:
        gameName:
          type: string
        platform:
          type: string
          enum: [PC, PlayStation, Xbox, "Nintendo Switch", "Steam Deck", iOS, Android]
        skillLevel:
          type: string
          enum: [Beginner, Intermediate, Expert, Chill]
        objectives:
          type: string
      required:
        - gameName
        - platform
        - skillLevel
        - objectives

    DashboardSummary:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        wallets:
          $ref: "#/components/schemas/Wallets"
        totalRequestsPosted:
          type: integer
        openRequestsCount:
          type: integer
        recentRequests:
          type: array
          items:
            $ref: "#/components/schemas/GameRequest"
      required:
        - user
        - wallets
        - totalRequestsPosted
        - openRequestsCount
        - recentRequests