# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gamerbuddy API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const signupBodyPasswordMin = 6;

export const SignupBody = zod.object({
  name: zod.string(),
  email: zod.string().email(),
  password: zod.string().min(signupBodyPasswordMin),
  phone: zod.string(),
  officialId: zod.instanceof(File).optional(),
});

/**
 * @summary Login with email and password
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    phone: zod.string(),
    idVerified: zod.boolean(),
    createdAt: zod.coerce.date(),
  }),
  message: zod.string().optional(),
});

/**
 * @summary Logout current user
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get current authenticated user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  idVerified: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get current user's wallets
 */
export const GetWalletsResponse = zod.object({
  hiringBalance: zod.number(),
  earningsBalance: zod.number(),
  canWithdraw: zod.boolean(),
  canPostRequest: zod.boolean(),
});

/**
 * @summary Add money to hiring wallet (min 10.75, max 1000)
 */
export const depositHiringWalletBodyAmountMin = 10.75;
export const depositHiringWalletBodyAmountMax = 1000;

export const DepositHiringWalletBody = zod.object({
  amount: zod
    .number()
    .min(depositHiringWalletBodyAmountMin)
    .max(depositHiringWalletBodyAmountMax),
});

export const DepositHiringWalletResponse = zod.object({
  hiringBalance: zod.number(),
  earningsBalance: zod.number(),
  canWithdraw: zod.boolean(),
  canPostRequest: zod.boolean(),
});

/**
 * @summary Withdraw from earnings wallet (min balance 100)
 */
export const WithdrawEarningsWalletBody = zod.object({
  amount: zod.number(),
});

export const WithdrawEarningsWalletResponse = zod.object({
  hiringBalance: zod.number(),
  earningsBalance: zod.number(),
  canWithdraw: zod.boolean(),
  canPostRequest: zod.boolean(),
});

/**
 * @summary List all game requests
 */
export const ListRequestsQueryParams = zod.object({
  platform: zod.coerce.string().optional(),
  skillLevel: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
});

export const ListRequestsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  userName: zod.string().optional(),
  gameName: zod.string(),
  platform: zod.enum([
    "PC",
    "PlayStation",
    "Xbox",
    "Nintendo Switch",
    "Steam Deck",
    "iOS",
    "Android",
  ]),
  skillLevel: zod.enum(["Beginner", "Intermediate", "Expert", "Chill"]),
  objectives: zod.string(),
  status: zod.enum(["open", "in_progress", "completed", "cancelled"]),
  createdAt: zod.coerce.date(),
});
export const ListRequestsResponse = zod.array(ListRequestsResponseItem);

/**
 * @summary Post a new game request (requires 10.75 in hiring wallet)
 */
export const CreateRequestBody = zod.object({
  gameName: zod.string(),
  platform: zod.enum([
    "PC",
    "PlayStation",
    "Xbox",
    "Nintendo Switch",
    "Steam Deck",
    "iOS",
    "Android",
  ]),
  skillLevel: zod.enum(["Beginner", "Intermediate", "Expert", "Chill"]),
  objectives: zod.string(),
});

/**
 * @summary Get current user's posted requests
 */
export const GetMyRequestsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  userName: zod.string().optional(),
  gameName: zod.string(),
  platform: zod.enum([
    "PC",
    "PlayStation",
    "Xbox",
    "Nintendo Switch",
    "Steam Deck",
    "iOS",
    "Android",
  ]),
  skillLevel: zod.enum(["Beginner", "Intermediate", "Expert", "Chill"]),
  objectives: zod.string(),
  status: zod.enum(["open", "in_progress", "completed", "cancelled"]),
  createdAt: zod.coerce.date(),
});
export const GetMyRequestsResponse = zod.array(GetMyRequestsResponseItem);

/**
 * @summary Get a specific request by ID
 */
export const GetRequestParams = zod.object({
  id: zod.coerce.number(),
});

export const GetRequestResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  userName: zod.string().optional(),
  gameName: zod.string(),
  platform: zod.enum([
    "PC",
    "PlayStation",
    "Xbox",
    "Nintendo Switch",
    "Steam Deck",
    "iOS",
    "Android",
  ]),
  skillLevel: zod.enum(["Beginner", "Intermediate", "Expert", "Chill"]),
  objectives: zod.string(),
  status: zod.enum(["open", "in_progress", "completed", "cancelled"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get dashboard summary for current user
 */
export const GetDashboardSummaryResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    phone: zod.string(),
    idVerified: zod.boolean(),
    createdAt: zod.coerce.date(),
  }),
  wallets: zod.object({
    hiringBalance: zod.number(),
    earningsBalance: zod.number(),
    canWithdraw: zod.boolean(),
    canPostRequest: zod.boolean(),
  }),
  totalRequestsPosted: zod.number(),
  openRequestsCount: zod.number(),
  recentRequests: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.number(),
      userName: zod.string().optional(),
      gameName: zod.string(),
      platform: zod.enum([
        "PC",
        "PlayStation",
        "Xbox",
        "Nintendo Switch",
        "Steam Deck",
        "iOS",
        "Android",
      ]),
      skillLevel: zod.enum(["Beginner", "Intermediate", "Expert", "Chill"]),
      objectives: zod.string(),
      status: zod.enum(["open", "in_progress", "completed", "cancelled"]),
      createdAt: zod.coerce.date(),
    }),
  ),
});