# lib/api-zod/src/generated/types/createRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gamerbuddy API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRequestBodyPlatform } from "./createRequestBodyPlatform";
import type { CreateRequestBodySkillLevel } from "./createRequestBodySkillLevel";

export interface CreateRequestBody {
  gameName: string;
  platform: CreateRequestBodyPlatform;
  skillLevel: CreateRequestBodySkillLevel;
  objectives: string;
}