# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gamerbuddy API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameRequest } from "./gameRequest";
import type { User } from "./user";
import type { Wallets } from "./wallets";

export interface DashboardSummary {
  user: User;
  wallets: Wallets;
  totalRequestsPosted: number;
  openRequestsCount: number;
  recentRequests: GameRequest[];
}